split embeds
